move singing stuff to its own file
